Introduced FirebaseCloudMessaging interface
